requirements.txt
astroid==2.0.4
certifi==2018.10.15
chardet==3.0.4
defusedxml==0.5.0
Django==2.2.10
django-allauth==0.38.0
django-braces==1.13.0
django-ckeditor==5.6.1
django-cleanup==4.0.0
django-cors-headers==2.4.0
django-jet==1.0.8
django-js-asset==1.1.0
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.3
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
feedparser==5.2.1
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.1.0
Pillow==5.3.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
PyJWT==1.6.4
pylint==2.1.1
python3-openid==3.1.0
pytz==2018.5
requests==2.20.0
requests-oauthlib==1.0.0
rest-social-auth==1.4.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
typed-ast==1.1.0
urllib3==1.24
uWSGI==2.0.17.1
wrapt==1.10.11
django-filter==2.2.0
